part of best practice 2
I have gone through and worked on the script for best practice #2. I am not sure that this is currently working, as the output gives an error, not an integer. I have looked through the code a few times and cant seem to see the problem. I have run out of time and it is time for me to take Charlie for a walk. I will do this another time when I have more patience than I do now...